Replace resource edit modal with inline row editing
Sidesteps the radix Dialog + base-ui Combobox interaction that swallowed
mouse clicks on tag suggestions, and matches the snappier inline pattern
used by the radar Blips table.
